Fix for potential clash with django 3.1a1 get_default isinstance bug
I believe that model.__base__ returned Base and rendering Base failed 
while instanciating Base in the template. The call of 
Base.user.remote_field.model returned a string "Auth.user" while a class 
was expected. However, rendering Base in the templatetag resolved this 
issue. I have to check if this changes the rendering of string for other 
Bases.